decompose endpoint config store file and remove store provider
