+++ src/app/Authenticated.tsx
@@ -22,7 +22,6 @@ import PageNotFound from 'pages/error/PageNotFound';
 import Home from 'pages/Home';
 import Materializations from 'pages/Materializations';
 import { Route, Routes } from 'react-router';
-import { EndpointConfigProvider } from 'stores/EndpointConfig';
 import { isProduction } from 'utils/env-utils';
 import { authenticatedRoutes, unauthenticatedRoutes } from './routes';
 
@@ -76,9 +75,7 @@ const Authenticated = () => {
                             element={
                                 <EntityContextProvider value="capture">
                                     <WorkflowContextProvider value="capture_create">
-                                        <EndpointConfigProvider>
-                                            <CaptureCreate />
-                                        </EndpointConfigProvider>
+                                        <CaptureCreate />
                                     </WorkflowContextProvider>
                                 </EntityContextProvider>
                             }
@@ -89,11 +86,9 @@ const Authenticated = () => {
                             element={
                                 <EntityContextProvider value="capture">
                                     <WorkflowContextProvider value="capture_edit">
-                                        <EndpointConfigProvider>
-                                            <EntityExistenceGuard>
-                                                <CaptureEdit />
-                                            </EntityExistenceGuard>
-                                        </EndpointConfigProvider>
+                                        <EntityExistenceGuard>
+                                            <CaptureEdit />
+                                        </EntityExistenceGuard>
                                     </WorkflowContextProvider>
                                 </EntityContextProvider>
                             }
@@ -117,9 +112,7 @@ const Authenticated = () => {
                             element={
                                 <EntityContextProvider value="materialization">
                                     <WorkflowContextProvider value="materialization_create">
-                                        <EndpointConfigProvider>
-                                            <MaterializationCreate />
-                                        </EndpointConfigProvider>
+                                        <MaterializationCreate />
                                     </WorkflowContextProvider>
                                 </EntityContextProvider>
                             }
@@ -132,11 +125,9 @@ const Authenticated = () => {
                             element={
                                 <EntityContextProvider value="materialization">
                                     <WorkflowContextProvider value="materialization_edit">
-                                        <EndpointConfigProvider>
-                                            <EntityExistenceGuard>
-                                                <MaterializationEdit />
-                                            </EntityExistenceGuard>
-                                        </EndpointConfigProvider>
+                                        <EntityExistenceGuard>
+                                            <MaterializationEdit />
+                                        </EntityExistenceGuard>
                                     </WorkflowContextProvider>
                                 </EntityContextProvider>
                             }
