+++ src/components/capture/Create.tsx
@@ -28,7 +28,8 @@ import {
     useDetailsForm_errorsExist,
     useDetailsForm_resetState,
 } from 'stores/DetailsForm';
-import { useEndpointConfigStore_reset } from 'stores/EndpointConfig';
+import { useEndpointConfigStore_reset } from 'stores/EndpointConfig/hooks';
+import { EndpointConfigHydrator } from 'stores/EndpointConfig/Hydrator';
 import {
     useFormStateStore_exitWhenLogsClose,
     useFormStateStore_resetState,
@@ -153,64 +154,70 @@ function CaptureCreate() {
                     'https://docs.estuary.dev/guides/create-dataflow/#create-a-capture',
             }}
         >
-            <ResourceConfigHydrator>
-                <ExistingEntityHydrator>
-                    <EntityCreate
-                        title="browserTitle.captureCreate"
-                        entityType={entityType}
-                        draftSpecMetadata={draftSpecsMetadata}
-                        resetState={resetState}
-                        errorSummary={
-                            <ValidationErrorSummary
-                                errorsExist={detailsFormErrorsExist}
-                            />
-                        }
-                        toolbar={
-                            <EntityToolbar
-                                GenerateButton={
-                                    <CaptureGenerateButton
-                                        entityType={entityType}
-                                        disabled={!hasConnectors}
-                                        callFailed={helpers.callFailed}
-                                        postGenerateMutate={mutateDraftSpecs}
-                                        createWorkflowMetadata={{
-                                            initiateDiscovery,
-                                            setInitiateDiscovery,
-                                        }}
-                                    />
-                                }
-                                TestButton={
-                                    <EntityTestButton
-                                        closeLogs={handlers.closeLogs}
-                                        callFailed={helpers.callFailed}
-                                        disabled={!hasConnectors}
-                                        logEvent={CustomEvents.CAPTURE_TEST}
-                                    />
-                                }
-                                SaveButton={
-                                    <EntitySaveButton
-                                        closeLogs={handlers.closeLogs}
-                                        callFailed={helpers.callFailed}
-                                        disabled={!draftId}
-                                        materialize={
-                                            handlers.materializeCollections
-                                        }
-                                        logEvent={CustomEvents.CAPTURE_CREATE}
-                                    />
-                                }
-                            />
-                        }
-                        RediscoverButton={
-                            <RediscoverButton
-                                entityType={entityType}
-                                disabled={!hasConnectors}
-                                callFailed={helpers.callFailed}
-                                postGenerateMutate={mutateDraftSpecs}
-                            />
-                        }
-                    />
-                </ExistingEntityHydrator>
-            </ResourceConfigHydrator>
+            <EndpointConfigHydrator>
+                <ResourceConfigHydrator>
+                    <ExistingEntityHydrator>
+                        <EntityCreate
+                            title="browserTitle.captureCreate"
+                            entityType={entityType}
+                            draftSpecMetadata={draftSpecsMetadata}
+                            resetState={resetState}
+                            errorSummary={
+                                <ValidationErrorSummary
+                                    errorsExist={detailsFormErrorsExist}
+                                />
+                            }
+                            toolbar={
+                                <EntityToolbar
+                                    GenerateButton={
+                                        <CaptureGenerateButton
+                                            entityType={entityType}
+                                            disabled={!hasConnectors}
+                                            callFailed={helpers.callFailed}
+                                            postGenerateMutate={
+                                                mutateDraftSpecs
+                                            }
+                                            createWorkflowMetadata={{
+                                                initiateDiscovery,
+                                                setInitiateDiscovery,
+                                            }}
+                                        />
+                                    }
+                                    TestButton={
+                                        <EntityTestButton
+                                            closeLogs={handlers.closeLogs}
+                                            callFailed={helpers.callFailed}
+                                            disabled={!hasConnectors}
+                                            logEvent={CustomEvents.CAPTURE_TEST}
+                                        />
+                                    }
+                                    SaveButton={
+                                        <EntitySaveButton
+                                            closeLogs={handlers.closeLogs}
+                                            callFailed={helpers.callFailed}
+                                            disabled={!draftId}
+                                            materialize={
+                                                handlers.materializeCollections
+                                            }
+                                            logEvent={
+                                                CustomEvents.CAPTURE_CREATE
+                                            }
+                                        />
+                                    }
+                                />
+                            }
+                            RediscoverButton={
+                                <RediscoverButton
+                                    entityType={entityType}
+                                    disabled={!hasConnectors}
+                                    callFailed={helpers.callFailed}
+                                    postGenerateMutate={mutateDraftSpecs}
+                                />
+                            }
+                        />
+                    </ExistingEntityHydrator>
+                </ResourceConfigHydrator>
+            </EndpointConfigHydrator>
         </PageContainer>
     );
 }
