+++ src/components/materialization/Create.tsx
@@ -24,7 +24,8 @@ import {
     useDetailsForm_errorsExist,
     useDetailsForm_resetState,
 } from 'stores/DetailsForm';
-import { useEndpointConfigStore_reset } from 'stores/EndpointConfig';
+import { useEndpointConfigStore_reset } from 'stores/EndpointConfig/hooks';
+import { EndpointConfigHydrator } from 'stores/EndpointConfig/Hydrator';
 import {
     useFormStateStore_exitWhenLogsClose,
     useFormStateStore_resetState,
@@ -132,51 +133,53 @@ function MaterializationCreate() {
                     'https://docs.estuary.dev/guides/create-dataflow/#create-a-materialization',
             }}
         >
-            <ResourceConfigHydrator>
-                <ExistingEntityHydrator>
-                    <EntityCreate
-                        title="browserTitle.materializationCreate"
-                        entityType={entityType}
-                        draftSpecMetadata={draftSpecsMetadata}
-                        resetState={resetState}
-                        errorSummary={
-                            <ValidationErrorSummary
-                                errorsExist={detailsFormErrorsExist}
-                            />
-                        }
-                        toolbar={
-                            <EntityToolbar
-                                GenerateButton={
-                                    <MaterializeGenerateButton
-                                        disabled={!hasConnectors}
-                                        callFailed={helpers.callFailed}
-                                    />
-                                }
-                                TestButton={
-                                    <EntityTestButton
-                                        disabled={!hasConnectors}
-                                        callFailed={helpers.callFailed}
-                                        closeLogs={handlers.closeLogs}
-                                        logEvent={
-                                            CustomEvents.MATERIALIZATION_TEST
-                                        }
-                                    />
-                                }
-                                SaveButton={
-                                    <EntitySaveButton
-                                        disabled={!draftId}
-                                        callFailed={helpers.callFailed}
-                                        closeLogs={handlers.closeLogs}
-                                        logEvent={
-                                            CustomEvents.MATERIALIZATION_CREATE
-                                        }
-                                    />
-                                }
-                            />
-                        }
-                    />
-                </ExistingEntityHydrator>
-            </ResourceConfigHydrator>
+            <EndpointConfigHydrator>
+                <ResourceConfigHydrator>
+                    <ExistingEntityHydrator>
+                        <EntityCreate
+                            title="browserTitle.materializationCreate"
+                            entityType={entityType}
+                            draftSpecMetadata={draftSpecsMetadata}
+                            resetState={resetState}
+                            errorSummary={
+                                <ValidationErrorSummary
+                                    errorsExist={detailsFormErrorsExist}
+                                />
+                            }
+                            toolbar={
+                                <EntityToolbar
+                                    GenerateButton={
+                                        <MaterializeGenerateButton
+                                            disabled={!hasConnectors}
+                                            callFailed={helpers.callFailed}
+                                        />
+                                    }
+                                    TestButton={
+                                        <EntityTestButton
+                                            disabled={!hasConnectors}
+                                            callFailed={helpers.callFailed}
+                                            closeLogs={handlers.closeLogs}
+                                            logEvent={
+                                                CustomEvents.MATERIALIZATION_TEST
+                                            }
+                                        />
+                                    }
+                                    SaveButton={
+                                        <EntitySaveButton
+                                            disabled={!draftId}
+                                            callFailed={helpers.callFailed}
+                                            closeLogs={handlers.closeLogs}
+                                            logEvent={
+                                                CustomEvents.MATERIALIZATION_CREATE
+                                            }
+                                        />
+                                    }
+                                />
+                            }
+                        />
+                    </ExistingEntityHydrator>
+                </ResourceConfigHydrator>
+            </EndpointConfigHydrator>
         </PageContainer>
     );
 }
